# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Tic-Tac-Toe API",
    "version": "0.1.0"
  },
  "paths": {
    "/games": {
      "get": {
        "tags": [
          "games"
        ],
        "summary": "List Games",
        "operationId": "list_games_games_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/GameResponse"
                  },
                  "type": "array",
                  "title": "Response List Games Games Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "games"
        ],
        "summary": "Create Game",
        "operationId": "create_game_games_post",
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GameResponse"
                }
              }
            }
          }
        }
      }
    },
    "/games/{game_id}": {
      "get": {
        "tags": [
          "games"
        ],
        "summary": "Get Game",
        "operationId": "get_game_games__game_id__get",
        "parameters": [
          {
            "name": "game_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Game Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GameResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "games"
        ],
        "summary": "Delete Game",
        "operationId": "delete_game_games__game_id__delete",
        "parameters": [
          {
            "name": "game_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Game Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/games/{game_id}/moves": {
      "post": {
        "tags": [
          "games"
        ],
        "summary": "Make Move",
        "operationId": "make_move_games__game_id__moves_post",
        "parameters": [
          {
            "name": "game_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Game Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MoveRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MoveResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "games"
        ],
        "summary": "List Game Moves",
        "operationId": "list_game_moves_games__game_id__moves_get",
        "parameters": [
          {
            "name": "game_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Game Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MoveHistoryItem"
                  },
                  "title": "Response List Game Moves Games  Game Id  Moves Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/games/{game_id}/coach": {
      "post": {
        "tags": [
          "games"
        ],
        "summary": "Coach Game",
        "operationId": "coach_game_games__game_id__coach_post",
        "parameters": [
          {
            "name": "game_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Game Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AICoachResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/index.html": {
      "get": {
        "summary": "Serve Index",
        "operationId": "serve_index_index_html_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/": {
      "get": {
        "summary": "Serve Index",
        "operationId": "serve_index__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AICoachResponse": {
        "properties": {
          "game": {
            "$ref": "#/components/schemas/GameResponse"
          },
          "recommended_position": {
            "type": "integer",
            "title": "Recommended Position"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "game",
          "recommended_position",
          "message"
        ],
        "title": "AICoachResponse"
      },
      "Board": {
        "properties": {
          "cells": {
            "items": {
              "type": "string",
              "enum": [
                "",
                "X",
                "O"
              ]
            },
            "type": "array",
            "title": "Cells"
          }
        },
        "type": "object",
        "required": [
          "cells"
        ],
        "title": "Board",
        "description": "Tic-tac-toe board: 9 cells, each \"\" or \"X\" or \"O\"."
      },
      "GameResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "board": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "current_player": {
            "type": "string",
            "title": "Current Player"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "winner": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Winner"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "board",
          "current_player",
          "status",
          "winner"
        ],
        "title": "GameResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "MoveHistoryItem": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "game_id": {
            "type": "string",
            "title": "Game Id"
          },
          "move_number": {
            "type": "integer",
            "title": "Move Number"
          },
          "player": {
            "type": "string",
            "enum": [
              "X",
              "O"
            ],
            "title": "Player"
          },
          "position": {
            "type": "integer",
            "title": "Position"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "game_id",
          "move_number",
          "player",
          "position",
          "created_at"
        ],
        "title": "MoveHistoryItem"
      },
      "MoveRequest": {
        "properties": {
          "x": {
            "type": "integer",
            "title": "X"
          },
          "y": {
            "type": "integer",
            "title": "Y"
          }
        },
        "type": "object",
        "required": [
          "x",
          "y"
        ],
        "title": "MoveRequest",
        "description": "Request payload for making the next move.\n\nCoordinates are 0-based, where (0, 0) is the top-left corner and (2, 2) is\nthe bottom-right corner of the 3\u00d73 board."
      },
      "MoveResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "board": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "current_player": {
            "type": "string",
            "title": "Current Player"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "winner": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Winner"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "id",
          "board",
          "current_player",
          "status",
          "winner"
        ],
        "title": "MoveResponse"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}